% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_1.R
\docType{data}
\name{poly_1}
\alias{poly_1}
\title{The First of Two Overlapping Polygon Layers for Demonstration Purposes}
\format{
Attributes:
\describe{
\item{A}{character}
\item{B}{integer}
\item{C}{year}
}
}
\usage{
poly_1
}
\description{
This polygon layer (\code{poly_1}) overlaps spatially with an other one
\code{\link{poly_2}}. They have one homonymous attribute \code{A} in common
and each of them is composed of four adjacent squares. These layers serve
the purpose of demonstrating the workings of the function \code{\link{st_or}}.
}
\keyword{datasets}
