% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_2.R
\docType{data}
\name{poly_2}
\alias{poly_2}
\title{The Second of Two Overlapping Polygon Layers for Demonstration Purposes}
\format{
Attributes:
\describe{
\item{A}{character}
\item{D}{integer}
\item{E}{year}
}
}
\usage{
poly_2
}
\description{
This polygon layer (\code{poly_2}) overlaps spatially with an other one
\code{\link{poly_1}}. They have one homonymous attribute \code{A} in common
and each of them is composed of four adjacent squares. These layers serve the
purpose of demonstrating the workings of the function \code{\link{st_or}}.
}
\keyword{datasets}
