% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfhelpers-package.R
\docType{package}
\name{sfhelpers-package}
\alias{sfhelpers}
\alias{sfhelpers-package}
\title{sfhelpers: Helper Functions for Wrapping Simple Features-Related Routines}
\description{
Functions helping to run routines related primarily to simple features (sf), but also to spatial objects of other classes / packages (terra, stars).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/a-benini/sfhelpers}
  \item \url{https://a-benini.github.io/sfhelpers/}
  \item Report bugs at \url{https://github.com/a-benini/sfhelpers/issues}
}

}
\author{
\strong{Maintainer}: Attilio Benini \email{attilio.benini@gmx.net}

}
\keyword{internal}
