% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_cell_size.R
\name{st_cell_size}
\alias{st_cell_size}
\title{Return cell size of a \code{SpatRaster} or \code{stars} object}
\usage{
st_cell_size(x, warn = TRUE, drop_units = FALSE)
}
\arguments{
\item{x}{object of class \code{SpatRaster}, \code{stars} or \code{stars_proxy}}

\item{warn}{logical; if \code{TRUE} (default) warn if \code{x} has no CRS or
geographic coordinates}

\item{drop_units}{logical; if \code{TRUE} returns output without units,
if \code{FALSE} (default) returns output with units if \code{x} has projected coordinates}
}
\value{
A single numeric value representing the cell size of input \code{x}
equivalent to the product of the input's x and y raster resolution. By default
and if \code{x} has projected coordinates, the cell size is returned with
the units according to input \code{x}'s CRS.
}
\description{
Return cell size of a \code{SpatRaster} or \code{stars} object
}
\examples{
library(terra)
library(sf)

# a raster with projected coordinates
r_projected <- rast(
  nrows = 10, ncols = 10, xmin = 0, xmax = 100, ymin = 0, ymax = 100,
  crs = "EPSG:2026"
)
values(r_projected) <- sample(5, ncell(r_projected), replace = TRUE)
# plot(r_projected)

# units of projected coordinates
st_crs(r_projected)$units
# pixel resolution of raster
res(r_projected)
# cell size of raster-pixel
st_cell_size(r_projected)
# cell size without units
st_cell_size(r_projected, drop_units = TRUE)

# a raster with geographic coordinates
r_geographic <- rast(
  nrows = 10, ncols = 10, xmin = 0, xmax = 100, ymin = 0, ymax = 100,
  crs = "EPSG:4326"
)
values(r_geographic) <- sample(5, ncell(r_geographic), replace = TRUE)
# plot(r_geographic)

# unavailable units of geographic coordinates
st_crs(r_geographic)$units
# pixel resolution of raster
res(r_geographic)
# if raster has geographic coordinates
# cell size is return by default without units and with a warning
st_cell_size(r_geographic)
# warning can be avoided
st_cell_size(r_geographic, warn = FALSE)
}
