% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_layers_as_df.R
\name{st_layers_as_df}
\alias{st_layers_as_df}
\title{Return properties of layers in a datasource as \code{data.frame}}
\usage{
st_layers_as_df(x, ..., drop_driver = TRUE, warn = TRUE)
}
\arguments{
\item{x}{any input that works for \code{\link[sf]{st_layers}}'s argument
\code{dsn} or an object of class \code{sf_layers}}

\item{...}{arguments of the function \code{\link[sf]{st_layers}} except
\code{dsn} if \code{x} is equivalent to \code{dsn} else ignored}

\item{drop_driver}{logical; \code{TRUE} (default) drops information about
the driver, \code{FALSE} adds the driver as column to the returned
\code{data.frame}}

\item{warn}{logical; \code{TRUE} (default) warns if no layer is available,
\code{FALSE} does not warn in such a case}
}
\value{
if no layer is available \code{NULL}, else a \code{data.frame}
resembling the \code{print}ed return of \code{\link[sf]{st_layers}} with
columns
\describe{
\item{layer_name}{name of the layer}
\item{geometry_type or geometry_types}{for each layer a string listing its
geometry type(s)}
\item{features}{number of features (if reported; see argument
\code{do_count} of \code{\link[sf]{st_layers}})}
\item{fields}{number of fields}
\item{crs}{short name of each layer's \code{crs}}
\item{driver (optional)}{name of driver, if argument \code{drop_driver} has
been set to \code{FALSE}}
}
}
\description{
Return properties of layers in a datasource as \code{data.frame}
}
\examples{
library(sf)
dsn <- system.file("shape", package = "sf")
st_layers_as_df(x = dsn, drop_driver = FALSE)

# properties of layers of multiple datasources in one data.frame
some_dsn <- list.files(system.file("gpkg", package = "sf"), full.names = TRUE)
multiple_dsn <- c(dsn, some_dsn)
df <-
  as.list(multiple_dsn) \%>\%
  setNames(multiple_dsn) \%>\%
  lapply(st_layers_as_df, drop_driver = FALSE) \%>\%
  data.table::rbindlist(use.names = FALSE, idcol = "dsn")
}
