% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_rename_geometry.R
\name{st_rename_geometry}
\alias{st_rename_geometry}
\title{Rename geometry column of an sf object}
\usage{
st_rename_geometry(obj, geometry_name)
}
\arguments{
\item{obj}{object of class \code{sf}}

\item{geometry_name}{a single character string renaming the active list-column with simple feature geometries}
}
\value{
a \code{sf} object with renamed geometry list-column.
}
\description{
Rename geometry column of an sf object
}
\details{
\code{st_rename_geometry()} is inspired by code found on
\href{https://gis.stackexchange.com/questions/386584/sf-geometry-column-naming-differences-r}{gis.stackexchange}
and particularly useful to homogenize \code{sf} objects with differently
named geometry columns or \code{list}s of such before binding them to a single
\code{sf} object by
\itemize{
\item \code{rbind} (\code{\link[sf]{bind}})
\item \code{do.call(rbind, <list_of_sf>)}
\item \code{sf::st_as_sf(data.table::rbindlist(<list_of_sf>))}
}
}
\examples{
# current active list-column with simple feature geometries:
attr(poly_1, "sf_column")
st_rename_geometry(poly_1, "renamed_geometry")
}
